+++ crates/ptool-lua/src/lua_api.rs
@@ -38,6 +38,8 @@ fn create_ptool_module(
     let use_fn = lua.create_function(move |_, required_version: String| {
         use_state.borrow().require_version(required_version)
     })?;
+    let version_state = Rc::clone(&world);
+    let version_fn = lua.create_function(move |_, ()| Ok(version_state.borrow().version()))?;
     let cd_state = Rc::clone(&world);
     let cd_fn = lua.create_function(move |_, path: String| cd_state.borrow_mut().cd(path))?;
     let unindent_state = Rc::clone(&world);
@@ -84,6 +86,7 @@ fn create_ptool_module(
     module.set("exec", exec_fn)?;
     module.set("config", config_fn)?;
     module.set("use", use_fn)?;
+    module.set("version", version_fn)?;
     module.set("cd", cd_fn)?;
     module.set("unindent", unindent_fn)?;
     module.set("inspect", inspect_fn)?;